requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
yt-dlp>=2025.8.22
requests==2.31.0
python-dotenv==1.0.0
google-api-python-client==2.107.0
Pillow==10.0.1
gunicorn
python-socks
beautifulsoup4
selenium>=4.15.0
supabase==2.0.0
